stockpilot/morning_report/requirements.txt
requests
python-dotenv
pandas
numpy
beautifulsoup4
lxml
